Work Completed on 12-14-2018
Work Completed on 12-14-2018 - Mainly worked on introducing the ability for the user to save their License Key, and to prevent the plugin from loading if the user hasn't entered a license key, or if it failed to validate. Everything is ready for creating the actual API call to validate a license, once EDD stuff is bought. One Day. Eventually.